pytest_django_crud/requirements.txt
pytest-django
pytest-yaml
loguru
pytest-html
pytest-cov
coverage